Refacto tri par status et staffing needed date
