+++ client/test/unit/specs/utils/jobsSorter.spec.js
@@ -1,23 +1,139 @@
 import projectStatus from '@/utils/projectStatus';
+import projectStaffingNeededDate from "@/utils/projectStaffingNeededDate";
+import jobsSorter from '@/utils/jobsSorter';
 
-describe('Unit | Utils | Project Status', () => {
-  it('should sort jobs according to project status', () => {
-    // Given
-    const job1 = { id: 1, project: { status: 'proposal_in_progress' } };
-    const job2 = { id: 2, project: { status: 'mission_signed' } };
-    const job3 = { id: 3, project: { status: 'mission_accepted' } };
-    const job4 = { id: 4, project: { status: 'proposal_in_progress' } };
-    const job5 = { id: 5, project: { status: 'proposal_sent' } };
-    const job6 = { id: 6, project: { status: 'lead' } };
+describe('Unit | Utils | Jobs Sorter', () => {
+  let jobs;
+  let expectedJobsWhenSortedByStatus;
+  let expectedJobsWhenSortedByStatusAndStaffingNeededDate;
+  beforeEach(() => {
+    // given
+    jobs = [
+      {
+        id: 1,
+        activity: {
+          title: 'Tech Lead mission 1',
+          staffing_needed_from: '2017-10-01',
+        },
+        project: {
+          id: 123456,
+          status: 'proposal_in_progress',
+          name: 'SCLOU - Cloud computing : enjeux, architecture et gouvernance du IaaS, CaaS, PaaS INTER 2017',
+          customer: {
+            name: 'La Poste - Courrier',
+          },
+          duration: '10 mois',
+          location: 'OCTO',
+          business_contact: {
+            nickname: 'ABC',
+          },
+          mission_director: {
+            nickname: 'XYZ',
+          },
+        },
+      },
+      {
+        id: 2,
+        activity: {
+          title: 'Tech Lead mission 2',
+          staffing_needed_from: '2017-10-02',
+        },
+        project: {
+          id: 123456,
+          status: 'mission_signed',
+          name: 'SCLOU - Cloud computing : enjeux, architecture et gouvernance du IaaS, CaaS, PaaS INTER 2017',
+          customer: {
+            name: 'La Poste - Courrier',
+          },
+          duration: '10 mois',
+          location: 'OCTO',
+          business_contact: {
+            nickname: 'ABC',
+          },
+          mission_director: {
+            nickname: 'XYZ',
+          },
+        },
+      },
+      {
+        id: 3,
+        activity: {
+          title: 'Tech Lead mission 3',
+          staffing_needed_from: '2017-10-03',
+        },
+        project: {
+          id: 123456,
+          status: 'proposal_in_progress',
+          name: 'SCLOU - Cloud computing : enjeux, architecture et gouvernance du IaaS, CaaS, PaaS INTER 2017',
+          customer: {
+            name: 'La Poste - Courrier',
+          },
+          duration: '10 mois',
+          location: 'OCTO',
+          business_contact: {
+            nickname: 'ABC',
+          },
+          mission_director: {
+            nickname: 'XYZ',
+          },
+        },
+      },
+      {
+        id: 4,
+        activity: {
+          title: 'Tech Lead mission 4',
+          staffing_needed_from: '2017-10-04',
+        },
+        project: {
+          id: 123456,
+          status: 'mission_signed',
+          name: 'SCLOU - Cloud computing : enjeux, architecture et gouvernance du IaaS, CaaS, PaaS INTER 2017',
+          customer: {
+            name: 'La Poste - Courrier',
+          },
+          duration: '10 mois',
+          location: 'OCTO',
+          business_contact: {
+            nickname: 'ABC',
+          },
+          mission_director: {
+            nickname: 'XYZ',
+          },
+        },
+      },
+    ];
+    expectedJobsWhenSortedByStatus = [
+      jobs[1],
+      jobs[3],
+      jobs[0],
+      jobs[2],
+    ];
+    expectedJobsWhenSortedByStatusAndStaffingNeededDate = [
+      jobs[3],
+      jobs[1],
+      jobs[2],
+      jobs[0],
+    ];
 
-    const givenJobs = [job1, job2, job3, job4, job5, job6];
-    const expectedJobs = [job2, job3, job5, job1, job4, job6];
+    sinon.stub(projectStatus, 'sort').returns(expectedJobsWhenSortedByStatus);
+    let staffingSort = sinon.stub(projectStaffingNeededDate, 'sort');
+    staffingSort.onCall(0)
+      .returns(expectedJobsWhenSortedByStatusAndStaffingNeededDate.slice(0,2));
+    staffingSort.onCall(1)
+      .returns(expectedJobsWhenSortedByStatusAndStaffingNeededDate.slice(2));
+  });
 
+  afterEach(() => {
+    projectStatus.sort.restore();
+    projectStaffingNeededDate.sort.restore();
+  });
+
+  it('should sort jobs by status and by staffing needed date', () => {
     // When
-    const sortedJobs = projectStatus.sort(givenJobs);
+    const sortedJobs = jobsSorter.sort(jobs);
 
     // Then
-    expect(sortedJobs).to.deep.equal(expectedJobs);
+    expect(sortedJobs).to.deep.equal(expectedJobsWhenSortedByStatusAndStaffingNeededDate);
   });
-});
 
+});
